Fix hidden definitions generation for Ruby 3.0
Use ruby2_keywords to support Ruby 2.6, 2.7, _and_ 3.0.

See the "compatible delegation" section in this blog post for more
information about this solution:
https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/

Essentially, Ruby changed the way it handles arguments in method
definitions, and this was causing deprecation warnings in Ruby 2.7.2,
which became an error with the release of Ruby 3.

I don't love that this defines ruby2_keywords in the main namespace like
this, so I'd be open to feedback as to how this
should be implemented. We could use the ruby2_keywords gem,
but I'm hesitant to add an extra depdendency to Sorbet.